requirements.txt
pandas==1.5.1
numpy==1.23.4
matplotlib==3.6.2
scikit-learn==1.1.3
streamlit-option-menu==0.3.2
torchtext==0.17.1
torch==2.2.1
tensorflow==2.10.1
Pillow==9.3.0
keras==2.10.0
tqdm==4.64.1